changed file names, add central page Home.tsx to place components
