+++ frontend/src/components/DateTime.tsx
@@ -1,4 +1,4 @@
-import React, {useState, useEffect} from 'react'
+import {useState, useEffect} from 'react'
 
 export const DateTime = () => {
     var [date, setDate] = useState(new Date());
@@ -16,5 +16,4 @@ export const DateTime = () => {
         </div>
     )
 }
-export{}
-export default DateTime
+
